Everything looks better
Added icons to select a module instead of text buttons, there is a logo
on the home page, module select has a reset bomb button which returns
you to home page.  Keypad now darkens and undarkens icons when clicked
and won’t let you click more than 4.  This does not actually work with
the logic, but at least it looks nice.  NEED TO FIX.  Rewrote password
to be much simpler and able to find a password with less input.
Renamed some icons in keypad while rewriting it, refracting it really
messed with some modules but I think I fixed them all (if you see
random ‘key’ around this is why).  Added an app icon image that needs
to be resized and placed in appropriate folders.  Renamed the app to
Keep Tapping and Nobody Explodes.